Fixes import errors created from renamings
