+++ src/pyrite/_camera/camera.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 from ..services import CameraService
 from ..enum import Layer
 from ..events import OnEnable, OnDisable
-from .._rendering import CameraRenderer
+from .._rendering.camera_renderer import CameraRendererProvider as CameraRenderer
 from .._rendering.viewport import Viewport
 from ..transform import transform_component
 from ..types import Camera as CameraBase, Renderable
