+++ src/pyrite/game.py
@@ -14,7 +14,8 @@ from .core.rate_settings import RateSettings
 from .core.system_manager import SystemManager
 
 from pyrite._camera.camera import Camera
-from ._rendering import OrthoProjection, Viewport
+from pyrite._rendering.ortho_projection import OrthoProjection
+from pyrite._rendering.viewport import Viewport
 from .services import CameraService
 from .transform import transform_system
 from .utils import threading
