+++ tests/test_render_manager.py
@@ -13,7 +13,7 @@ from pyrite.core.render_system import (
 )
 from pyrite._entity.entity import BaseEntity
 from pyrite.enum import Layer, RenderLayers
-from pyrite._rendering import OrthoProjection, RectBounds
+from pyrite.rendering import OrthoProjection, RectBounds
 from pyrite._rendering.base_renderable import BaseRenderable
 from pyrite.services import CameraService
 from pyrite.types import CullingBounds
